Move numericdate_conversion into separate file and make public
